add role menu-client and menu-admin
