+++ client/src/Components/navbar/navbar.jsx
@@ -27,6 +27,7 @@ function Navbars() {
   async function setClient() {
     try {
       const token = await getAccessTokenSilently();
+       console.log(token)
       const info = await axios.get("http://localhost:4000/login/setClient", {
         headers: {
           authorization: `Bearer ${token}`,
@@ -142,13 +143,7 @@ async function getInfo() {
           )}
     </> 
         </Container>
-      </Navbar>
-
-                 {
-                  isAuthenticated && 
-                  <pre style={{textAlign: 'start'}}>{JSON.stringify(user, null, 2)}</pre>
-               }
-          
+      </Navbar>         
     </>
   );
 }
